+++ ai_safety_gridworlds/helpers/gridworld_gym_env.py
@@ -155,7 +155,7 @@ class GridworldGymEnv(gym.Env):
     def get_episode_no(self):                           # ADDED
         return self._env.get_episode_no()
 
-    # gym does not support additional arguments to .step() method so we need to use a separate method
+    # gym does not support additional arguments to .step() method so we need to use a separate method. See also https://github.com/openai/gym/issues/2399
     def set_current_q_value_per_action(self, q_value_per_action):                           # ADDED
         return self._env.set_current_q_value_per_action(q_value_per_action)